PayOS integration
Add PayOS SDK + API

- feat: Add Payment API for Order/Stages
- refractor: money field double -> BigDecimal
- Importance: HAVE NOT TESTED API

+++ src/main/java/org/example/catholicsouvenircustomorder/dto/request/ProductCreateDTO.java
@@ -4,6 +4,8 @@ package org.example.catholicsouvenircustomorder.dto.request;
 import lombok.Data;
 import org.example.catholicsouvenircustomorder.model.OrderDetail;
 import org.example.catholicsouvenircustomorder.model.ProductImage;
+
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -11,7 +13,7 @@ import java.util.UUID;
 public class ProductCreateDTO {
     private UUID artisanId;
     private String productName;
-    private double productPrice;
+    private BigDecimal productPrice;
     private String productDescription;
     private int quantity;
     private boolean status;
